requirements.txt
langchain
langchain-groq
langchain-community
langgraph
pydantic
pydantic-settings
python-dotenv
streamlit
pandas
google-genai
tavily-python
fastapi
uvicorn
